+++ internal/models/modulerelease.go
@@ -15,9 +15,9 @@ type ModuleRelease struct {
 	StatusID    uint                `json:"status_id"`
 	Status      ModuleReleaseStatus `gorm:"foreignKey:StatusID"`
 	Keywords    []Keyword           `gorm:"many2many:release_keywords;" json:"keywords,omitempty"`
-	DiskSize    uint                `json:"disk_size"`
+	DiskSize    int64               `json:"disk_size"`
 }
 
-func NewModuleReleaseFromModule(m *Module, version string, status ModuleReleaseStatus, diskSize uint) *ModuleRelease {
+func NewModuleReleaseFromModule(m *Module, version string, status ModuleReleaseStatus, diskSize int64) *ModuleRelease {
 	return &ModuleRelease{ModuleID: m.ID, Version: version, Description: m.Description, StatusID: status.ID, Status: status, DiskSize: diskSize}
 }
